(feat)graph: implement deep first search

+++ graph/include/graph.hpp
@@ -17,44 +17,44 @@ using namespace algs4;
 namespace algs4 {
     class Graph {
     private:
-        unsigned long v = 0;
-        unsigned long e = 0;
-        vector<vector<unsigned long>> adj;
+        int v = 0;
+        int e = 0;
+        vector<vector<int>> adj;
 
     public:
-        explicit Graph(const unsigned long &pV) : v(pV), e(0), adj(v) {}
+        explicit Graph(const int &pV) : v(pV), e(0), adj(v) {}
 
         explicit Graph(InStream &pIn) {
             v = pIn.ReadInt();
-            unsigned long edge = pIn.ReadInt();
+            int edge = pIn.ReadInt();
             adj.assign(v, {});
-            for (unsigned long index = 0; index < edge; index++) {
-                unsigned long v = pIn.ReadInt();
-                unsigned long w = pIn.ReadInt();
+            for (int index = 0; index < edge; index++) {
+                int v = pIn.ReadInt();
+                int w = pIn.ReadInt();
                 AddEdge(v, w);
             }
         }
 
-        unsigned long V() { return v; }
+        int V() { return v; }
 
-        unsigned long E() { return e; }
+        int E() { return e; }
 
-        void AddEdge(unsigned long v, unsigned long w) {
+        void AddEdge(int v, int w) {
             adj[v].push_back(w);
             adj[w].push_back(v);
             e++;
         }
 
-        vector<unsigned long> ADJ(unsigned long v) {
+        vector<int> ADJ(int v) {
             return adj[v];
         }
 
         string ToString() {
             ostringstream os;
             os << "Graph: " << v << " vertices, " << e << " edges\n";
-            for (unsigned long index = 0; index < v; index++) {
+            for (int index = 0; index < v; index++) {
                 os << index << ": ";
-                for (unsigned long w:ADJ(index)) {
+                for (int w:ADJ(index)) {
                     os << w << " ";
                 }
                 os << "\n";
