+++ graph/include/utils.hpp
@@ -11,7 +11,7 @@
 using namespace std;
 
 namespace algs4 {
-    pair<string, unsigned long> ParseCommandLine(int argc, const char **argv) {
+    pair<string, int> ParseCommandLine(int argc, const char **argv) {
         if (argc < 3) {
             cerr << "Usage: executor [test_file_path] [source_vertex]" << endl;
             exit(0);
@@ -23,7 +23,7 @@ namespace algs4 {
             cerr << "source should be greater than 0 (" << tempSource << ")" << endl;
             exit(0);
         }
-        unsigned long source = strtoul(argv[2], nullptr, 0);
+        int source = strtoul(argv[2], nullptr, 0);
 
         return {testFile, source};
     };
